Fix issue when musicians id is hypenated
If the musician id is hypenated and the conductor shares a name, i.e. `todo` and `todo-item` then `_searchForConductor` would fail as the query would return the musician itself rather than the controller.

This fixes that by looking for the conductor from the musician from its parent up rather than itself up (as `closest` includes the element itself in the search).

This change also minor re-factoring to define the keys used for the conductor and musician controller properties in only one place for consistency.